+++ src/api/productService.js
@@ -2,7 +2,7 @@ import { uploadProductImage } from '../utils/uploadProductImage';
 import { supabase } from './client';
 
 /**
- * productService
+ * addProduct
  * @description 상품 등록 서비스
  * @param {object} product - 상품 데이터
  * @param {string} userId - 사용자 ID
@@ -27,6 +27,14 @@ export const addProduct = async (product, userId) => {
   return { data, error };
 };
 
+/**
+ * updateProduct
+ * @description 상품 등록 서비스
+ * @param {object} product
+ * @param {string} userId
+ * @param {number} productId
+ * @returns
+ */
 export const updateProduct = async (product, userId, productId) => {
   // 이미지 업로드
   const file = product.img;
@@ -53,4 +61,21 @@ export const updateProduct = async (product, userId, productId) => {
 
   if (error) throw error;
   return { data, error };
+};
+
+/**
+ * getProductDetail
+ * @description 상품 상세 데이터 조회
+ * @param {number} productId
+ * @returns
+ */
+export const getProductDetail = async (productId) => {
+  const { data, error } = await supabase
+    .from('products')
+    .select('*')
+    .eq('id', productId)
+    .single();
+
+  if (error) throw error;
+  return data;
 };
